requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-core
faiss-cpu
pypdf
pandas
requests
python-dotenv
fastapi
uvicorn
python-multipart